Create form-section and center it
Used flexbox on the container holding both the photo and form sections to get them
aligned and centered into middle of page. This took a lot of playing around with margins
and the justify-content property. Aslo replaced the method Iwas using to pull the image
in. Instead of background-image, <img> element was used. Additionally, styled the
background of the <body element>