begin styling of the form-section
All three divs contained in this section will stretch the entire lengh of the screen
vertically. Working to get the measurements of each div and their margins and paddings
worked out. Making sure to use methods that leave the lengths of the gaps on each
edge of the screen unchanged