Develop <form> structure
Beginning the process of implementing the form into the html. The form will ask the user/
player's last team whereabouts, their personal information, and what team they desire
to play for. The form will make up the middle section of the form section, and will be
the largest in size