Further style input and select fields
More focus was given on the phone-number input section. A pattern and maxlength attribute
were added to require users to type in a valid US phone number. A placeholder was also
inserted to give them guidance. Input and select diths were also expanded, along with
increasing font size of the input elements