Fill out .title-section
Introduction to the league and and some persuasion text has beenplaced here to get
users to sign up. The position of the the league's logo was also adjusted. Still working to
condense the hook and foloowing sentences properly to avoid overflow into the form-section,
as well as overflowing apst the borders of the .title-section div